==> 2. Pre_Processamento/0. Função_Churn.ipynb <==
import pandas as pd

def calculate_churn(df):
    df = df.sort_values(by=['user_id', 'transaction_date'])
    df['next_transaction_date'] = df.groupby('user_id')['transaction_date'].shift(-1)
    df['next_expiration_date'] = df.groupby('user_id')['membership_expiration_date'].shift(-1)

    # Se não houver nova transação dentro de 30 dias após a expiração atual, conta como churn
    df['churn'] = (df['next_transaction_date'].isna()) | (
        (df['next_transaction_date'] - df['membership_expiration_date']).dt.days > 30)

    return df

# df = calculate_churn(df)
# print(df[['user_id', 'transaction_date', 'membership_expiration_date', 'churn']])
