==> 2. Pre_Processamento/1. Calculo_Churn_Amostragem.ipynb <==
import pandas as pd
import os
import datetime as dt

from google.colab import drive
drive.mount('/content/drive')

# Definindo os caminhos
folder_input = "/content/drive/Shareddrives/Real_Project_Churn_DSA30/Dataset/Raw/kkboxs"
folder_output = "/content/drive/Shareddrives/Real_Project_Churn_DSA30/Dataset/Bronze"

# Leitura Arquivos
df_members = pd.read_csv(os.path.join(folder_input, 'members_v3.csv'))
df_transactions = pd.read_csv(os.path.join(folder_input, 'transactions.csv'))

### Amostragem Aleatória (Base de Membros)

df_members = df_members[~df_members['gender'].isnull()]

df_members_semple_1 = df_members.sample(frac = 0.1, random_state = 21)
df_members_semple_2 = df_members.sample(frac = 0.1, random_state = 10)
df_members_semple_3 = df_members.sample(frac = 0.1, random_state = 42)


tipo_amostragem = "Amostragem_Aleatória"

print(df_members_semple_1.shape)
print(df_members_semple_2.shape)
print(df_members_semple_3.shape)
df_members_semple_1.to_csv(os.path.join(folder_output, 'Amostra_Clientes', tipo_amostragem, 'members_sample_1.csv'), index = False)
df_members_semple_2.to_csv(os.path.join(folder_output, 'Amostra_Clientes', tipo_amostragem,'members_sample_2.csv'), index = False)
df_members_semple_3.to_csv(os.path.join(folder_output, 'Amostra_Clientes', tipo_amostragem,'members_sample_3.csv'), index = False)

### Validação base de Transações

df_transactions_members = pd.merge(df_transactions, df_members, on='msno', how='inner')

df_transactions_members = df_transactions_members[df_transactions.columns]

df_transactions_members.head()

### Calculando Churn

time_churn = 30

def calculate_churn(df):
    df = df.sort_values(by=['msno', 'transaction_date', 'membership_expire_date'])
    df['next_transaction_date'] = df.groupby('msno')['transaction_date'].shift(-1)
    df['next_expiration_date'] = df.groupby('msno')['membership_expire_date'].shift(-1)

    df['churn'] = (df['next_transaction_date'].isna()) | (
        (df['next_transaction_date'] - df['membership_expire_date']).dt.days > time_churn)

    return df

df_transactions_members['membership_expire_date'] = pd.to_datetime(df_transactions_members['membership_expire_date'].astype(str), format='%Y%m%d')
df_transactions_members['transaction_date'] = pd.to_datetime(df_transactions_members['transaction_date'].astype(str), format='%Y%m%d')

df_transactions_members['transaction_date'].max()

data_base = df_transactions_members['transaction_date'].max() - dt.timedelta(days=time_churn)

data_base

df_churn = calculate_churn(df_transactions_members)

df_transactions_churn = df_churn[df_churn['membership_expire_date'] < data_base].drop(columns=['next_transaction_date', 'next_expiration_date'])

df_transactions_churn.head()

df_transactions_churn.to_csv(os.path.join(folder_input, 'df_transactions_churn_full.csv'), index=False)

### Amostragem Estratificada

df_clientes_churn = df_transactions_churn.groupby('msno')['churn'].max().reset_index()

df_clientes_churn['churn'].value_counts(normalize=True)

tipo_amostragem = "Amostragem_Estratificada"
from sklearn.model_selection import StratifiedShuffleSplit

i = 0
for ramdom_state in [21, 10, 42]:
  i += 1
  print(f"Rodada: {i}")
  split = StratifiedShuffleSplit(n_splits=1, test_size=0.8, random_state=ramdom_state)

  for train_idx, _ in split.split(df_clientes_churn, df_clientes_churn['churn']):
      amostra = df_clientes_churn.iloc[train_idx]
      amostra.to_csv(os.path.join(folder_output, 'Amostra_Clientes', tipo_amostragem, f'churn_sample_{i}.csv'), index=False)

  print(f"Tamanho da amostra{i} estratificada: {len(amostra)}")
