==> 2. Pre_Processamento/2. Amostragem_Bases.ipynb <==
import pandas as pd
import os

from google.colab import drive
drive.mount('/content/drive')

folder_raw = '/content/drive/Shareddrives/Real_Project_Churn_DSA30/Dataset/Raw/kkboxs'
folder_bronze = '/content/drive/Shareddrives/Real_Project_Churn_DSA30/Dataset/Bronze'

# Leitura Arquivos Raw
df_transaction = pd.read_csv(os.path.join(folder_raw, 'transactions.csv'))
df_members = pd.read_csv(os.path.join(folder_raw, 'members_v3.csv'))
df_transaction_churn = pd.read_csv(os.path.join(folder_raw, 'df_transactions_churn_full.csv'))

#Escolher amostra. (aleatória ou estratificada)

df_sample = pd.read_csv(os.path.join(folder_bronze, 'Amostra_Clientes', 'Amostragem_Estratificada' , 'churn_sample_1.csv'), usecols=['msno'])

df_transactions_sample = pd.merge(df_sample, df_transaction, on='msno', how='inner')
df_transactions_churn_sample = pd.merge(df_sample, df_transaction_churn, on='msno', how='inner')
df_members_sample = pd.merge(df_sample, df_members, on='msno', how='inner')

print('Numero de linhas da transactions: ', df_transactions_sample.shape[0])
print('Numero de linhas da transactions churn: ', df_transactions_churn_sample.shape[0])
print('Numero de linhas da members: ', df_members_sample.shape[0])

df_transactions_sample.to_csv(os.path.join(folder_bronze, 'df_transactions.csv'), index=False)
df_transactions_churn_sample.to_csv(os.path.join(folder_bronze, 'df_transactions_churn.csv'), index=False)
df_members_sample.to_csv(os.path.join(folder_bronze, 'df_members.csv'), index=False)

i = 0
for arquivo in os.listdir(os.path.join(folder_raw, 'user_logs')):
  if arquivo.endswith('.csv'):
    i+=1
    df_log = pd.read_csv(os.path.join(folder_raw, 'user_logs', arquivo))
    df_log_sample = pd.merge(df_sample, df_log, on='msno', how='inner')
    df_log_sample.to_csv(os.path.join(folder_bronze, 'User_Logs_Amostrados', f'df_userlog_{i}.csv'), index=False)

df_userlogs_1 = pd.DataFrame()
df_userlogs_2 = pd.DataFrame()

i = 0
user_logs = []

for arquivo in os.listdir(os.path.join(folder_bronze, 'User_Logs_Amostrados')):
  if i <= 20:
    i+=1
    user_logs.append(arquivo)
    print(arquivo)
    df_log = pd.read_csv(os.path.join(folder_bronze, 'User_Logs_Amostrados', arquivo))
    df_userlogs_1 = pd.concat([df_userlogs_1, df_log])


for arquivo in os.listdir(os.path.join(folder_bronze, 'User_Logs_Amostrados')):
  if arquivo not in user_logs:
    print(arquivo)
    df_log = pd.read_csv(os.path.join(folder_bronze, 'User_Logs_Amostrados', arquivo))
    df_userlogs_2 = pd.concat([df_userlogs_2, df_log])



df_userlogs_1.to_csv(os.path.join(folder_bronze, 'df_userlogs_1.csv'), index=False)
df_userlogs_2.to_csv(os.path.join(folder_bronze, 'df_userlogs_2.csv'), index=False)
