==> 2. Pre_Processamento/3. Agrupa_Trans_Churn_Mês.ipynb <==
import pandas as pd
import os
import datetime as dt

from google.colab import drive
drive.mount('/content/drive')

folder_bronze = '/content/drive/Shareddrives/Real_Project_Churn_DSA30/Dataset/Bronze'

df_transactions_churn = pd.read_csv(os.path.join(folder_bronze, 'df_transactions_churn.csv'))

df_transactions_churn['ano_mes'] = pd.to_datetime(df_transactions_churn['membership_expire_date']).dt.strftime('%Y-%m')

df_ult_expire_mes = df_transactions_churn.groupby(['msno', 'ano_mes'])[['churn', 'membership_expire_date']].max().reset_index()

df_churn_mes = pd.merge(
                        df_ult_expire_mes[['msno', 'membership_expire_date', 'churn']],
                        df_transactions_churn.drop_duplicates(['msno', 'membership_expire_date']),
                        on=['msno', 'membership_expire_date'],
                        how='left'
                        )
df_churn_mes.drop(columns=['churn_y'], inplace = True)
df_churn_mes.rename(columns={'churn_x': 'churn'}, inplace=True)

df_churn_mes['dt_ref'] = pd.to_datetime(df_churn_mes['ano_mes']) + pd.DateOffset(months=1)

df_churn_mes.to_csv(os.path.join(folder_bronze, 'df_transactions_churn_mes.csv'), index=False)
